ENH: RollDay `uadd` and refactors (#18)
(cherry picked from commit d652e6c15d1053b2275cb65880ae3a0f63e94903)
